fabs instead of fabsf

+++ MDSDK/MDSDK/Classes/ColorPickerView/UIColor+Equal.m
@@ -15,7 +15,7 @@
     CGFloat r1, g1, b1, a1, r2, g2, b2, a2;
     [self getRed:&r1 green:&g1 blue:&b1 alpha:&a1];
     [other getRed:&r2 green:&g2 blue:&b2 alpha:&a2];
-    return fabsf(r1 - r2) <= tolerance && fabsf(g1 - g2) <= tolerance && fabsf(b1 - b2) <= tolerance && fabsf(a1 - a2) <= tolerance;
+    return fabs(r1 - r2) <= tolerance && fabs(g1 - g2) <= tolerance && fabs(b1 - b2) <= tolerance && fabs(a1 - a2) <= tolerance;
 }
 
 @end
